Add impl From<(f64, f64)> for Point; re-export OrderedFloat type
Since a Point is made of OrderedFloat types, and it comes from another
crate, the only way for a user of this library to create a Point is by
adding a dependency on ordered_float in their own project, causing
unnecessary dependency bloat.

This allows users to use the OrderedFloat type used in this crate, or
create a Point using a tuple of floats.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>